+++ packages/core/src/connectivity/scp-cf/destination/destination-accessor-authentication-type.spec.ts
@@ -103,7 +103,7 @@ describe('authentication types', () => {
       httpMocks.forEach(mock => expect(mock.isDone()).toBe(true));
     });
 
-    it('should use provider token first instead of the userJwt when SystemUser exists in destination', async () => {
+    it('should use provider client credentials token for SystemUser exists in provider destination', async () => {
       mockServiceBindings();
       mockVerifyJwt();
       mockServiceToken();
@@ -120,6 +120,7 @@ describe('authentication types', () => {
           providerServiceToken
         ),
         mockSubaccountDestinationsCall(nock, [], 200, providerServiceToken),
+        // This single destination call is the one triggered by the OAuth2SAMLBearerAssertion flow
         mockSingleDestinationCall(
           nock,
           oauthSingleResponse,
@@ -132,6 +133,42 @@ describe('authentication types', () => {
       const expected = parseDestination(oauthSingleResponse);
       const actual = await getDestination(destinationName, {
         cacheVerificationKeys: false
+      });
+      expect(actual).toMatchObject(expected);
+      httpMocks.forEach(mock => expect(mock.isDone()).toBe(true));
+    });
+
+    it('should use subscriber client credentials token for SystemUser exists in subscriber destination', async () => {
+      mockServiceBindings();
+      mockVerifyJwt();
+      mockServiceToken();
+
+      const samlDestinationsWithSystemUser = { ...oauthMultipleResponse[0] };
+      // Insert SystemUser in the retrieved OAuth2SAMLBearer destination to trigger principle propagation workflow
+      samlDestinationsWithSystemUser['SystemUser'] = 'defined';
+
+      const httpMocks = [
+        mockInstanceDestinationsCall(
+          nock,
+          [samlDestinationsWithSystemUser],
+          200,
+          subscriberServiceToken
+        ),
+        mockSubaccountDestinationsCall(nock, [], 200, subscriberServiceToken),
+        // This single destination call is the one triggered by the OAuth2SAMLBearerAssertion flow
+        mockSingleDestinationCall(
+          nock,
+          oauthSingleResponse,
+          200,
+          destinationName,
+          subscriberServiceToken
+        )
+      ];
+
+      const expected = parseDestination(oauthSingleResponse);
+      const actual = await getDestination(destinationName, {
+        cacheVerificationKeys: false,
+        userJwt: subscriberUserJwt
       });
       expect(actual).toMatchObject(expected);
       httpMocks.forEach(mock => expect(mock.isDone()).toBe(true));